backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai>=2.16.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.16.2",
    "supabase>=2.27.3",
    "uvicorn>=0.40.0",
    "woocommerce>=3.0.0",
]
